# Allow global ErrorHandler to be set multiple times