Resolves #2140 

These changes allow a user to set the global `ErrorHandler` multiple times. This is done by storing the state in an `atomic.Value`.

This also adds a benchmark to track performance and overhead of these changes. The increase in memory usage and allocations appears related to the setting operation. Prior to the changes setting a global `ErrorHandler` had no memory overhead and with these changes it requires an allocation. This seems acceptable given this code correctly allows multiple setting of an `ErrorHandler` and this operation is expected to be called a limited amount of times during initialization of user code.

### Benchmark Before

```sh
$ go test -timeout 300s -bench=.
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
BenchmarkErrorHandler-8                  	  542686	      2256 ns/op	     144 B/op	       6 allocs/op
BenchmarkGetDefaultErrorHandler-8        	1000000000	         0.7989 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetDelegatedErrorHandler-8      	1000000000	         0.9686 ns/op	       0 B/op	       0 allocs/op
BenchmarkDefaultErrorHandlerHandle-8     	 2890158	       415.4 ns/op	      48 B/op	       1 allocs/op
BenchmarkDelegatedErrorHandlerHandle-8   	 4011504	       303.8 ns/op	      48 B/op	       1 allocs/op
BenchmarkSetErrorHandlerDelegation-8     	 7772974	       158.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkSetErrorHandlerNoDelegation-8   	344574063	         3.454 ns/op	       0 B/op	       0 allocs/op
```

### Benchmark After

```sh
$ go test -timeout=300s -bench=.
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
BenchmarkErrorHandler-8                  	  480949	      2306 ns/op	     176 B/op	       8 allocs/op
BenchmarkGetDefaultErrorHandler-8        	660670377	         1.887 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetDelegatedErrorHandler-8      	627903168	         1.903 ns/op	       0 B/op	       0 allocs/op
BenchmarkDefaultErrorHandlerHandle-8     	 4298862	       288.5 ns/op	      48 B/op	       1 allocs/op
BenchmarkDelegatedErrorHandlerHandle-8   	 3892299	       301.8 ns/op	      48 B/op	       1 allocs/op
BenchmarkSetErrorHandlerDelegation-8     	 3984223	       342.1 ns/op	      16 B/op	       1 allocs/op
BenchmarkSetErrorHandlerNoDelegation-8   	25017502	        46.58 ns/op	      16 B/op	       1 allocs/op
```